# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PromptFolio API (served by the Python FastAPI backend)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [catalog]
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [catalog]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    delete:
      operationId: deleteCategory
      tags: [catalog]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # ---------------- PROMPTS (CATALOG) ----------------
  /prompts:
    get:
      operationId: listPrompts
      tags: [catalog]
      parameters:
        - { name: search, in: query, required: false, schema: { type: string } }
        - { name: categoryId, in: query, required: false, schema: { type: string } }
        - { name: modality, in: query, required: false, schema: { $ref: "#/components/schemas/Modality" } }
        - { name: tag, in: query, required: false, schema: { type: string } }
        - { name: folder, in: query, required: false, schema: { type: string } }
        - { name: publishedOnly, in: query, required: false, schema: { type: boolean } }
      responses:
        "200":
          description: Prompts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prompt"
...
  /prompts/export:
    get:
      operationId: exportPrompts
      tags: [catalog]
      parameters:
        - name: format
          in: query
          required: true
          schema:
            type: string
            enum: [json, markdown]
      responses:
        "200":
          description: Export bundle (file content as text)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportBundle"
  /prompts/import:
    post:
      operationId: importPrompts
      tags: [catalog]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /prompts/suggest-variables:
    post:
      operationId: suggestPromptVariables
      tags: [catalog]
      summary: AI-written purpose, examples and required flag for every variable of a template (not saved)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuggestVariablesInput"
      responses:
        "200":
          description: Documented variables
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuggestVariablesResult"
...
  /prompts/{id}:
    get:
      operationId: getPrompt
      tags: [catalog]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
    patch:
      operationId: updatePrompt
      tags: [catalog]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromptUpdate"
      responses:
        "200":
          description: Updated (version is incremented when the template changes)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
    delete:
      operationId: deletePrompt
      tags: [catalog]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # ---------------- AGENTS ----------------
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }

    # ---------- enums ----------
    Modality:
      type: string
      enum: [image, video, audio, text]
...
    Category:
      type: object
      required: [id, name, slug, color, modalityScope, promptCount]
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        color: { type: string }
        modalityScope:
          type: string
          enum: [all, image, video, audio, text]
        promptCount: { type: integer }
    CategoryInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        color: { type: string }
        modalityScope:
          type: string
          enum: [all, image, video, audio, text]

    # ---------- prompts ----------
    PromptVariable:
      type: object
      required: [name]
      properties:
        name: { type: string }
        label: { type: ["string", "null"] }
        defaultValue: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        helpText: { type: ["string", "null"] }
        examples:
          type: array
          items: { type: string }
        options:
          type: array
          items: { type: string }
        required: { type: boolean }
    SuggestVariablesInput:
      type: object
      required: [promptTemplate]
      properties:
        promptTemplate: { type: string }
        variables:
          type: array
          items: { $ref: "#/components/schemas/PromptVariable" }
    SuggestVariablesResult:
      type: object
      required: [variables]
      properties:
        variables:
          type: array
          items: { $ref: "#/components/schemas/PromptVariable" }
    AssetThumb:
      type: object
      required: [id, modality, previewUrl]
      properties:
        id: { type: string }
        modality: { $ref: "#/components/schemas/Modality" }
        previewUrl: { type: string }
    Prompt:
      type: object
      required: [id, title, promptTemplate, variables, tags, targetModality, isPublished, version, recentAssets, assetCount, createdAt, updatedAt]
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: ["string", "null"] }
        promptTemplate: { type: string }
        variables:
          type: array
          items: { $ref: "#/components/schemas/PromptVariable" }
        tags:
          type: array
          items: { type: string }
        folder: { type: ["string", "null"] }
        categoryId: { type: ["string", "null"] }
        categoryName: { type: ["string", "null"] }
        categoryColor: { type: ["string", "null"] }
        targetModality: { $ref: "#/components/schemas/Modality" }
        preferredMcpId: { type: ["string", "null"] }
        preferredMcpName: { type: ["string", "null"] }
        sourceAgentName: { type: ["string", "null"] }
        isPublished: { type: boolean }
        version: { type: integer }
        recentAssets:
          type: array
          items: { $ref: "#/components/schemas/AssetThumb" }
        assetCount: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    PromptUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        description: { type: ["string", "null"] }
        promptTemplate: { type: string, minLength: 1 }
        variables:
          type: array
          items: { $ref: "#/components/schemas/PromptVariable" }
        tags:
          type: array
          items: { type: string }
        folder: { type: ["string", "null"] }
        categoryId: { type: ["string", "null"] }
        targetModality: { $ref: "#/components/schemas/Modality" }
        preferredMcpId: { type: ["string", "null"] }
        isPublished: { type: boolean }
...
    ExportBundle:
      type: object
      required: [filename, mimeType, content, count]
      properties:
        filename: { type: string }
        mimeType: { type: string }
        content: { type: string }
        count: { type: integer }
    ImportInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          description: JSON text previously produced by the JSON export
    ImportResult:
      type: object
      required: [imported, skipped]
      properties:
        imported: { type: integer }
        skipped: { type: integer }
...